fix compile

+++ src/server/game/Battlegrounds/Arena.cpp
@@ -27,8 +27,7 @@
 
 void ArenaScore::AppendToPacket(WorldPacket& data)
 {
-    data << uint64(PlayerGuid);
-
+    data << PlayerGuid;
     data << uint32(KillingBlows);
     data << uint8(TeamId);
     data << uint32(DamageDone);
